Compiled FDTD Method
Just the diagram on page 96 compromised of some of the work we have done so far. Is not operational as of yet and I would appreciate if anyone is able to make sense of how to organise these files in swift, I have handled most of the errors and only formatting left to do.